Fix #370 - Scala.js support is broken
Add
- `scalajs-java-securerandom` to use `java.security.SecureRandom` for `java.util.UUID`
- `scala-java-time` for `java.time` for Scala.js
- a custom URL for Scala.js because there's no alternative to `java.net.URL`

Also make some necessary changes including removing the code that won't be available in JavaScript.

Fix the tests for JavaScript. There are some issues with `Long` and `BigInt` less than -9007199254740991L and greater than 9007199254740991L.
