+++ modules/refined4s-circe/jvm/src/test/scala/refined4s/modules/circe/derivation/types/NumericTestValues.scala
@@ -1,0 +1,9 @@
+package refined4s.modules.circe.derivation.types
+
+/** @author Kevin Lee
+  * @since 2024-10-28
+  */
+trait NumericTestValues {
+  val MinLongValue: Long = Long.MinValue
+  val MaxLongValue: Long = Long.MaxValue
+}
