+++ project/plugins.sbt
@@ -4,7 +4,7 @@ addSbtPlugin("org.wartremover" % "sbt-wartremover" % "3.2.0")
 
 addSbtPlugin("ch.epfl.scala" % "sbt-scalafix"  % "0.12.1")
 addSbtPlugin("org.scalameta" % "sbt-scalafmt"  % "2.5.2")
-addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.1.1")
+addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.2.2")
 addSbtPlugin("org.scalameta" % "sbt-mdoc"      % "2.5.4")
 addSbtPlugin("io.kevinlee"   % "sbt-docusaur"  % "0.16.0")
